✨feat: 測定結果をFirestoreに登録する処理を追加

+++ lib/app/application/usecase/training/training_usecase.dart
@@ -25,7 +25,7 @@ class TrainingUsecase with RunUsecaseMixin {
     required ResultRank rank,
     required int correct,
     required int questions,
-    required int correctRate,
+    required double correctRate,
     required DateTime doneAt,
   }) =>
       execute(
